Refactor: import문을 이용한 스크립트 모듈화
